requirements.txt
yt-dlp
discord.py
ffmpeg
pynacl
spotipy
